perform cleanup and refactoring
