+++ experiments/mnist/config_best.yaml
@@ -1,9 +1,9 @@
 ---
-__object__: laplace_flows.experiments.predefined.ExperimentCollection
+__object__: laplace_flows.experiments.base.ExperimentCollection
 name: mnist_basedist_comparison
 experiments:
     - &exp_laplace_best
-      __object__: laplace_flows.experiments.predefined.HyperoptExperiment
+      __object__: laplace_flows.experiments.hyperopt.HyperoptExperiment
       name: mnist_normal_best
       scheduler: &scheduler 
         __object__: ray.tune.schedulers.ASHAScheduler
