+++ experiments/mnist/config_lu.yaml
@@ -1,9 +1,9 @@
 ---
-__object__: veriflow.experiments.predefined.ExperimentCollection
+__object__: veriflow.experiments.baseExperimentCollection
 name: mnist_basedist_comparison
 experiments:
     - &exp_laplace
-      __object__: veriflow.experiments.predefined.HyperoptExperiment
+      __object__: veriflow.experiments.hyperopt.HyperoptExperiment
       name: mnist_laplace
       scheduler: &scheduler 
         __object__: ray.tune.schedulers.ASHAScheduler
