--- presentation/tfl-style.css
@@ -1,178 +1,0 @@
-.column {
-    float: left;
-    width: 50%;
-  }
-  
-/* Clear floats after the columns */
-.row:after {
-content: "";
-display: table;
-clear: both;
-}
-
-
-  /*
-This file is mainly copy-pasta from rise's examples
-https://github.com/damianavila/RISE/blob/master/examples/rise.css
-that was further customized for appliedAI purposes
-*/
-@import url('https://fonts.googleapis.com/css2?family=Work+Sans:wght@400&display=swap');
-
-
-body {
-    font-family: 'Work Sans', sans-serif !important;
-    text-transform: initial !important;
-    letter-spacing: initial !important;
-    font-weight: 400 !important;
-    line-height: 1.5 !important;
-    text-size-adjust: 100% !important;
-    ‑webkit‑text‑size‑adjust: 100% !important;
-}
-
-
-.reveal, .md-slide, .sidebar-wrapper {
-    font-size: 1.5rem !important;
-}
-
-.navbar-default .navbar-nav > li > a {
-    color: #00747b !important;
-}
-
-.filename {
-    font-size: 2.4rem !important;
-    color: #212529 !important;
-    font-weight: 600 !important;
-}
-
-div.text_cell_render {
-    font-size: 3rem !important;
-    line-height: 1.5 !important;
-}
-
-.reveal, .md-slide {
-    color: white !important;
-    font-size: 1.5rem !important;
-}
-
-h1 {
-    position:fixed;
-    top: 25px;
-    color: #00747b !important;
-    font-size: 200%;
-}
-h2 {
-    color: #00747b !important;
-}
-
-h3, h4, h5, h6 {
-    color: #808080 !important;
-}
-
-.reveal p, .reveal ol, .reveal dl, .reveal ul,
-div.text_cell_render {
-    color: #212529 !important;
-}
-
-/*copied from stackoverflow, better spacing between list items*/
-li + li {
-  margin-top: 0.2em;
-}
-
-body.rise-enabled .reveal ol, body.rise-enabled .reveal dl, body.rise-enabled .reveal ul {
-    margin-left: 0.1em;
-    margin-top: 0.2em;
-}
-
-.reveal .rendered_html h1:first-child,
-.reveal .rendered_html h2:first-child,
-.reveal .rendered_html h3:first-child,
-.reveal .rendered_html h4:first-child,
-.reveal .rendered_html h5:first-child {
-    margin-top: 0.2em;
-}
-
-.CodeMirror-lines, .output_text {
-    font-size: 1.5rem !important;
-}
-
-
-h1.plan, h2.plan, h3.plan {
-    text-align: center;
-    padding-bottom: 30px;
-}
-
-ul.plan>li>span.plan-bold {
-    font-size: 110%;
-    padding: 4px;
-    font-weight: bold;
-    background-color: #eee;
-}
-
-ul.plan>li>ul.subplan>li>span.plan-bold {
-    font-weight: bold;
-}
-
-.plan-strike {
-    opacity: 0.4;
-/*    text-decoration: line-through; */
-}
-
-div.plan-container {
-    display: grid;
-    grid-template-columns: 50% 50%;
-}
-
-/*
- * this is to void xarray's html output to show the fallback textual representation
- * see also
-   * xarray.md and
-   * https://github.com/damianavila/RISE/issues/594
- */
-.reveal pre.xr-text-repr-fallback {
-    display: none;
-}
-
-#toc-header, .toc-item li {
-    margin: auto !important;
-    color: #808080 !important;
-}
-
-#toc, #toc-wrapper, .toc-item-num, #toc a, .toc {
-    margin: auto !important;
-    color: #00747b !important;
-}
-
-#toc-wrapper {
-    top: auto !important;
-    bottom: auto !important;
-    margin-top: 2rem !important;
-    color: #00747b !important;
-}
-
-
-#rise-header {
-    margin: 10px;
-    left: 5%;
-}
-
-#rise-footer {
-    margin: 10px;
-    right: 5%;
-}
-
-#rise-backimage {
-    opacity: 0.70;
-}
-
-.reveal img {
-    max-width: 100%;
-    top: -50%;
-}
-
-
-.md-slide.title {
-  position: relative;
-  top: -50%;
-  margin-left: 5%;
-}
-
