--- presentation/tox.ini
@@ -1,26 +1,0 @@
-[tox]
-requires =
-    tox>=4
-env_list = lint, type, py{311}
-
-[testenv]
-description = run unit tests
-deps =
-    pytest>=7
-    pytest-sugar
-commands =
-    pytest {posargs:tests}
-
-[testenv:lint]
-description = run linters
-skip_install = true
-deps =
-    black==22.12
-commands = black {posargs:.}
-
-[testenv:type]
-description = run type checks
-deps =
-    mypy>=0.991
-commands =
-    mypy {posargs:src tests}
