+++ veriflow/experiments/base.py
@@ -1,8 +1,5 @@
-import typing as T
 import os
-import uuid
-import pandas as pd
-from datetime import datetime
+import typing as T
 
 
 class Experiment(object):
@@ -37,4 +34,31 @@ class Experiment(object):
         """Conducts the experiment and saves the results to the report directory. The method is expected to store all results in report_dir.
         """
         raise NotImplementedError
+    
+class ExperimentCollection(Experiment):
+    """ Implements an experiment that consists of several jointly conducted but independent experiments.
+    """
+    def __init__(self, experiments: T.Iterable[Experiment], *args, **kwargs) -> None:
+        """
+        The function initializes an object with a list of experiments based on a given configuration.
+        
+        :param experiments: The "experiments" parameter is an iterable object that contains a list of
+        experiments. Each experiment is represented by a configuration object
+        :type experiments: Iterable *args
+        """
+        super().__init__(*args, **kwargs)
+        self.experiments = experiments
+    
+    @classmethod
+    def from_dict(cls, config: T.Dict[str, T.Any]) -> "ExperimentCollection":
+        config = deepcopy(config)
+        for i, exp_cfg in enumerate(config["experiment_params"]["experiments"]):
+            config["experiment_params"]["experiments"][i] = Experiment.from_dict(exp_cfg)   
+        
+        return Experiment.from_dict(config)
+        
+    
+    def conduct(self, report_dir: os.PathLike, storage_path: os.PathLike = None):
+        for i, exp in enumerate(self.experiments):
+            exp.conduct(os.path.join(report_dir, f"{i}_{exp.name}"), storage_path=storage_path)
