--- veriflow/experiments/serialization.py
@@ -1,6 +1,0 @@
-import pickle 
-
-def unpickle(file):
-    with open(file, 'rb') as fo:
-        dict = pickle.load(fo, encoding='bytes')
-    return dict
